Refactored HighlightsProp to be placed closer to where it is used.
