+++ src/components/PlaybackBar/Playbackbar.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import Slider from './Slider';
 import { PlaybackStyle, PlaySlider } from './playbackStyle';
-import { HighlightsProp } from '../../types/global';
+import { HighlightsProp } from './playbackUtil';
 import { ThemeProvider } from '@emotion/react';
 import theme from 'src/styles/theme';
