+++ src/components/PlaybackBar/Slider.tsx
@@ -1,9 +1,8 @@
 import { FC, useMemo, useRef } from 'react';
 
-import { getLeft, getPercentage, getValue, throttled } from './playbackUtil';
+import { getLeft, getPercentage, getValue, throttled, HighlightsProp } from './playbackUtil';
 import { SliderContainer, SliderBar, SliderStyle, ThumbStyle } from './playbackStyle';
 import HighlightSections from './HighlightSections';
-import { HighlightsProp } from '../../types/global';
 
 interface SliderProps {
     value: number;
